requirements.txt
leandropls.easylambda==0.1.3
openai==1.58.1
pydantic==2.10.4
